tools/requirements.txt
python-whois>=0.9.6
shodan>=1.31.1
requests>=2.31.0
